backend/requirements.txt
fastapi
uvicorn[standard]

supabase
sqlalchemy
psycopg2-binary

python-dotenv

pandas
numpy
python-dateutil

scikit-learn
joblib

langchain
langchain-core
langchain-community
langgraph
google-generativeai

twilio

python-docx
reportlab

apscheduler

requests
aiohttp
pydantic
jinja2
orjson

loguru
